* Modify domain and data models * Add tests for remote layer

+++ remote/src/main/java/de/iskae/remote/model/SysModel.kt
@@ -2,7 +2,7 @@ package de.iskae.remote.model
 
 data class SysModel(
     val country: String,
-    val sunrise: Int,
-    val sunset: Int,
+    val sunrise: Long,
+    val sunset: Long,
     val type: Int
 )
